Add an example for installing java 8
Prior to this commit, the documentation was lacking an example of how to install a specific version of java.  Which led users to believe that they should use the version parameter.  

After this commit, we have an example showing the correct way to install a different major verison of java is to use the package parameter.