(538) - Unresolved fact fix
Prior to this commit, when run on non-darwin systems with no java
installed, the custom fact java_version was unresolved.

This was due to an incorrect return type set in
lib/facter/java_version.rb (returned a boolean).

This commit hopes to rectify this issue, by ensuring that the returned
value for java_version is in the correct format.