Fix containment of `java::config`
In 0dcd91b6e8d2b2519228c534fe4976e9d2a3a1d6 the `anchor` pattern was
removed, but not replaced with anything and thus broke the containment
of `java::config`.  This restores the original behaviour by using
`contain`.

As noted in the code comment, the declaration remains in the same place
in the code as evaluation order is important.  `java::config` is a
private class and relies on variables from the base class's scope having
been set before it is evaluated.